requirements.txt
python-telegram-bot==20.1
openai==0.27.2
